Demo Sidebar component created (#3)
* sidebar component on left

I just added an empty sidebar component on the left. the home.component is now split into a home.component__left class and home.component__right class using 'div' in html. The left side will be used for the sidebar and the right side will be used for the mxgraph editor.

* deleting vbox